Update Nova\Database and Nova\Modules

+++ src/Modules/Console/Commands/ModuleMigrateRefreshCommand.php
@@ -3,6 +3,7 @@
 namespace Nova\Modules\Console\Commands;
 
 use Nova\Console\Command;
+use Nova\Console\ConfirmableTrait;
 
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
@@ -31,6 +32,8 @@ class ModuleMigrateRefreshCommand extends Command
      */
     public function fire()
     {
+        if (! $this->confirmToProceed()) return;
+
         $slug = $this->argument('slug');
 
         $this->call('module:migrate:reset', array(
